Separated log targets into different files
