+++ core/include/mmcore/utility/log/Log.h
@@ -16,12 +16,6 @@
 #include <iostream>
 #include <thread>
 
-#define SPDLOG_EOL ""
-
-#include "spdlog/sinks/basic_file_sink.h"
-#include "spdlog/sinks/msvc_sink.h"
-#include "spdlog/sinks/ostream_sink.h"
-#include "spdlog/spdlog.h"
 
 namespace megamol {
 namespace core {
@@ -160,278 +154,6 @@ namespace log {
 
             /** The log level for this target */
             UINT level;
-
-        };
-
-#ifdef _WIN32
-        /**
-         * Target class echoing the log messages into a stream
-         */
-        class DebugOutputTarget : public Target {
-        public:
-
-            /**
-             * Ctor
-             *
-             * @param level The log level used for this target
-             */
-            DebugOutputTarget(UINT level = Log::LEVEL_ERROR);
-
-            /** Dtor */
-            virtual ~DebugOutputTarget(void);
-
-            /**
-             * Writes a message to the log target
-             *
-             * @param level The level of the message
-             * @param time The time stamp of the message
-             * @param sid The object id of the source of the message
-             * @param msg The message text itself
-             */
-            void Msg(UINT level, TimeStamp time, SourceID sid,
-                const char *msg) override;
-
-            /**
-             * Writes a message to the log target
-             *
-             * @param level The level of the message
-             * @param time The time stamp of the message
-             * @param sid The object id of the source of the message
-             * @param msg The message text itself
-             */
-            void Msg(UINT level, TimeStamp time, SourceID sid, std::string const& msg) override;
-
-        private:
-            std::shared_ptr<spdlog::sinks::msvc_sink_mt> sink;
-
-            std::shared_ptr<spdlog::logger> logger;
-        };
-#endif /* _WIN32 */
-
-        /**
-         * Target class safing message to a ASCII text file
-         */
-        class FileTarget : public Target {
-        public:
-
-            /**
-             * Opens a physical log file
-             *
-             * @param path The path to the physical log file
-             * @param level The log level used for this target
-             */
-            FileTarget(std::string const& path, UINT level = Log::LEVEL_ERROR);
-
-            /** Dtor */
-            virtual ~FileTarget(void);
-
-            /** Flushes any buffer */
-            void Flush(void) override;
-
-            /**
-             * Answer the path to the physical log file
-             *
-             * @return The path to the physical log file
-             */
-            inline const std::string& Filename(void) const {
-                return this->filename;
-            }
-
-            /**
-             * Writes a message to the log target
-             *
-             * @param level The level of the message
-             * @param time The time stamp of the message
-             * @param sid The object id of the source of the message
-             * @param msg The message text itself
-             */
-            void Msg(UINT level, TimeStamp time, SourceID sid,
-                const char *msg) override;
-
-            /**
-             * Writes a message to the log target
-             *
-             * @param level The level of the message
-             * @param time The time stamp of the message
-             * @param sid The object id of the source of the message
-             * @param msg The message text itself
-             */
-            void Msg(UINT level, TimeStamp time, SourceID sid, std::string const& msg) override;
-
-        private:
-            
-            /** The file name of the log file used */
-            std::string filename;
-
-            std::shared_ptr<spdlog::sinks::basic_file_sink_mt> sink;
-
-            std::shared_ptr<spdlog::logger> logger;
-
-        };
-
-        /**
-         * Target class storing message as long as the file is offline
-         */
-        class OfflineTarget : public Target {
-        public:
-
-            /**
-             * Ctor
-             *
-             * @param bufferSize The number of message to be stored in the
-             *                   offline buffer.
-             * @param level The log level used for this target
-             */
-            OfflineTarget(unsigned int bufferSize = 20,
-                UINT level = Log::LEVEL_ERROR);
-
-            /** Dtor */
-            virtual ~OfflineTarget(void);
-
-            /**
-             * Answer the size of the offline message buffer
-             *
-             * @return The number of offline message that will be stored in
-             *         the offline buffer
-             */
-            inline unsigned int BufferSize(void) const {
-                return this->bufSize;
-            }
-
-            /**
-             * Writes a message to the log target
-             *
-             * @param level The level of the message
-             * @param time The time stamp of the message
-             * @param sid The object id of the source of the message
-             * @param msg The message text itself
-             */
-            void Msg(UINT level, TimeStamp time, SourceID sid,
-                const char *msg) override;
-
-            /**
-             * Writes a message to the log target
-             *
-             * @param level The level of the message
-             * @param time The time stamp of the message
-             * @param sid The object id of the source of the message
-             * @param msg The message text itself
-             */
-            void Msg(UINT level, TimeStamp time, SourceID sid, std::string const& msg) override;
-
-            /**
-             * Answer the number of omitted messages
-             *
-             * @return The number of omitted messages
-             */
-            inline unsigned int OmittedMessagesCount(void) const {
-                return omittedCnt;
-            }
-
-            /**
-             * Echoes all stored offline messages to 'target' and optionally
-             * deletes the message buffer.
-             *
-             * @param target The target object to receive the messages
-             * @param remove If true the offline message buffer will be freed
-             */
-            void Reecho(Target &target, bool remove = true);
-
-            /**
-             * Sets the size of the buffer for offline messages
-             *
-             * @param bufferSize The number of message to be stored in the
-             *                   offline buffer.
-             */
-            void SetBufferSize(unsigned int bufferSize);
-
-        private:
-
-            /**
-             * Utility struct for offline messages
-            */
-            typedef struct _offlinemessage_t {
-
-                /** The level of the message */
-                UINT level;
-
-                /** The message time stamp */
-                TimeStamp time;
-
-                /** The message source ID */
-                SourceID sid;
-
-                /** The message text */
-                std::string msg;
-
-            } OfflineMessage;
-
-            /** The number of offline messages to be stored */
-            unsigned int bufSize;
-
-            /** The number of stored messages */
-            unsigned int msgCnt;
-
-            /** buffer of offline messages */
-            OfflineMessage *msgs;
-
-            /** The number of omitted messages */
-            unsigned int omittedCnt;
-
-        };
-
-        /**
-         * Target class echoing the log messages into a stream
-         */
-        class MEGAMOLCORE_API StreamTarget : public Target {
-        public:
-
-            /** Stream target to stdout */
-            static const std::shared_ptr<Target> StdOut;
-
-            /** Stream target to stderr */
-            static const std::shared_ptr<Target> StdErr;
-
-            /**
-             * Ctor
-             *
-             * @param stream The stream to write the log messages to
-             * @param level The log level used for this target
-             */
-            StreamTarget(std::ostream& stream, UINT level = Log::LEVEL_ERROR);
-
-            /** Dtor */
-            virtual ~StreamTarget(void);
-
-            /** Flushes any buffer */
-            void Flush(void) override;
-
-            /**
-             * Writes a message to the log target
-             *
-             * @param level The level of the message
-             * @param time The time stamp of the message
-             * @param sid The object id of the source of the message
-             * @param msg The message text itself
-             */
-            void Msg(UINT level, TimeStamp time, SourceID sid,
-                const char *msg) override;
-
-            /**
-             * Writes a message to the log target
-             *
-             * @param level The level of the message
-             * @param time The time stamp of the message
-             * @param sid The object id of the source of the message
-             * @param msg The message text itself
-             */
-            void Msg(UINT level, TimeStamp time, SourceID sid, std::string const& msg) override;
-
-        private:
-
-             std::shared_ptr<spdlog::sinks::ostream_sink_mt> sink;
-
-             std::shared_ptr<spdlog::logger> logger;
 
         };
