+++ core/src/utility/log/Log.cpp
@@ -23,6 +23,11 @@
 #include <sys/stat.h>
 #endif /* _WIN32 */
 
+#include "mmcore/utility/log/OfflineTarget.h"
+#include "mmcore/utility/log/FileTarget.h"
+
+#include "spdlog/spdlog.h"
+
 
 const char megamol::core::utility::log::Log::std_pattern[3] = "%v";
 
@@ -52,275 +57,6 @@ void megamol::core::utility::log::Log::Target::Flush(void) {
     // intentionally empty
 }
 
-/*****************************************************************************/
-#ifdef _WIN32
-
-/*
- * megamol::core::utility::log::Log::DebugOutputTarget::DebugOutputTarget
- */
-megamol::core::utility::log::Log::DebugOutputTarget::DebugOutputTarget(UINT level)
-        : Target(level) {
-    logger = spdlog::get("default_debug");
-    if (logger == nullptr) {
-        sink = std::make_shared<spdlog::sinks::msvc_sink_mt>();
-        sink->set_pattern(std_pattern);
-        logger = std::make_shared<spdlog::logger>("default_debug", sink);
-    }
-}
-
-
-/*
- * megamol::core::utility::log::Log::DebugOutputTarget::~DebugOutputTarget
- */
-megamol::core::utility::log::Log::DebugOutputTarget::~DebugOutputTarget(void) {
-    // intentionally empty
-}
-
-
-/*
- * megamol::core::utility::log::Log::DebugOutputTarget::Msg
- */
-void megamol::core::utility::log::Log::DebugOutputTarget::Msg(UINT level,
-        megamol::core::utility::log::Log::TimeStamp time, megamol::core::utility::log::Log::SourceID sid,
-        const char *msg) {
-    Msg(level, time, sid, std::string(msg));
-}
-
-
-/*
- * megamol::core::utility::log::Log::DebugOutputTarget::Msg
- */
-void megamol::core::utility::log::Log::DebugOutputTarget::Msg(
-    UINT level, TimeStamp time, SourceID sid, std::string const& msg) {
-    if (level > this->Level()) return;
-    logger->info("{}|{}", level, msg);
-}
-
-#endif /* _WIN32 */
-/*****************************************************************************/
-
-/*
- * megamol::core::utility::log::Log::FileTarget::FileTarget
- */
-megamol::core::utility::log::Log::FileTarget::FileTarget(std::string const& path, UINT level)
-        : Target(level) {
-    logger = spdlog::get(std::string("default_file_") + path);
-    if (logger == nullptr) {
-        sink = std::make_shared<spdlog::sinks::basic_file_sink_mt>(path, true);
-        sink->set_pattern(std_pattern);
-        logger = std::make_shared<spdlog::logger>(std::string("default_file_") + path, sink);
-    }
-    filename = path;
-}
-
-
-/*
- * megamol::core::utility::log::Log::FileTarget::~FileTarget
- */
-megamol::core::utility::log::Log::FileTarget::~FileTarget(void) {
-    logger->flush();
-}
-
-
-/*
- * megamol::core::utility::log::Log::FileTarget::Flush
- */
-void megamol::core::utility::log::Log::FileTarget::Flush(void) {
-    logger->flush();
-}
-
-
-/*
- * megamol::core::utility::log::Log::FileTarget::Msg
- */
-void megamol::core::utility::log::Log::FileTarget::Msg(UINT level,
-        megamol::core::utility::log::Log::TimeStamp time, megamol::core::utility::log::Log::SourceID sid,
-        const char *msg) {
-    Msg(level, time, sid, std::string(msg));
-}
-
-
-/*
- * megamol::core::utility::log::Log::FileTarget::Msg
- */
-void megamol::core::utility::log::Log::FileTarget::Msg(UINT level, megamol::core::utility::log::Log::TimeStamp time,
-    megamol::core::utility::log::Log::SourceID sid, std::string const& msg) {
-    if (level > this->Level()) return;
-
-    struct tm* timeStamp;
-#ifdef _WIN32
-#    if (_MSC_VER >= 1400)
-    struct tm __tS;
-    timeStamp = &__tS;
-    if (localtime_s(timeStamp, &time) != 0) {
-        // timestamp error
-        __tS.tm_hour = __tS.tm_min = __tS.tm_sec = 0;
-    }
-#    else  /* (_MSC_VER >= 1400) */
-    timeStamp = localtime(&time);
-#    endif /* (_MSC_VER >= 1400) */
-#else      /* _WIN32 */
-    timeStamp = localtime(&time);
-#endif     /* _WIN32 */
-    logger->info("{}:{}:{}|{}|{}|{}", timeStamp->tm_hour, timeStamp->tm_min, timeStamp->tm_sec, sid, level, msg);
-}
-
-/*****************************************************************************/
-
-/*
- * megamol::core::utility::log::Log::OfflineTarget::OfflineTarget
- */
-megamol::core::utility::log::Log::OfflineTarget::OfflineTarget(unsigned int bufferSize,
-        UINT level) : Target(level), bufSize(bufferSize), msgCnt(0),
-        msgs(new OfflineMessage[bufferSize]), omittedCnt(0) {
-    // intentionally empty
-}
-
-
-/*
- * megamol::core::utility::log::Log::OfflineTarget::~OfflineTarget
- */
-megamol::core::utility::log::Log::OfflineTarget::~OfflineTarget(void) {
-    ARY_SAFE_DELETE(this->msgs);
-    this->bufSize = 0;
-    this->msgCnt = 0;
-}
-
-
-/*
- * megamol::core::utility::log::Log::OfflineTarget::Msg
- */
-void megamol::core::utility::log::Log::OfflineTarget::Msg(UINT level,
-        megamol::core::utility::log::Log::TimeStamp time, megamol::core::utility::log::Log::SourceID sid,
-        const char *msg) {
-    Msg(level, time, sid, std::string(msg));
-}
-
-
-/*
- * megamol::core::utility::log::Log::OfflineTarget::Msg
- */
-void megamol::core::utility::log::Log::OfflineTarget::Msg(UINT level, megamol::core::utility::log::Log::TimeStamp time,
-    megamol::core::utility::log::Log::SourceID sid, std::string const& msg) {
-    // Do not check the level. We store ALL messages
-    if (this->msgCnt < this->bufSize) {
-        this->msgs[this->msgCnt].level = level;
-        this->msgs[this->msgCnt].time = time;
-        this->msgs[this->msgCnt].sid = sid;
-        this->msgs[this->msgCnt].msg = msg;
-        this->msgCnt++;
-    } else {
-        this->omittedCnt++;
-    }
-}
-
-
-/*
- * megamol::core::utility::log::Log::OfflineTarget::Reecho
- */
-void megamol::core::utility::log::Log::OfflineTarget::Reecho(megamol::core::utility::log::Log::Target &target,
-        bool remove) {
-    for (unsigned int i = 0; i < this->msgCnt; i++) {
-        target.Msg(this->msgs[i].level, this->msgs[i].time, this->msgs[i].sid,
-            this->msgs[i].msg);
-    }
-    if (remove) this->msgCnt = 0;
-    if (this->omittedCnt > 0) {
-        std::stringstream omg;
-        omg << this->omittedCnt << " offline log message" << ((this->omittedCnt == 1) ? "" : "s") << " omitted\n";
-        target.Msg(Log::LEVEL_WARN, Log::CurrentTimeStamp(),
-            Log::CurrentSourceID(), omg.str());
-        if (remove) this->omittedCnt = 0;
-    }
-}
-
-
-/*
- * megamol::core::utility::log::Log::OfflineTarget::SetBufferSize
- */
-void megamol::core::utility::log::Log::OfflineTarget::SetBufferSize(unsigned int bufferSize) {
-    OfflineMessage *om = this->msgs;
-    this->msgs = new OfflineMessage[bufferSize];
-    unsigned int cnt = std::min(this->msgCnt, bufferSize);
-    this->omittedCnt += (this->msgCnt - cnt);
-    this->bufSize = bufferSize;
-    for (unsigned int i = 0; i < cnt; i++) {
-        this->msgs[i].level = om[i].level;
-        this->msgs[i].time = om[i].time;
-        this->msgs[i].sid = om[i].sid;
-        this->msgs[i].msg = om[i].msg;
-    }
-    delete[] om;
-}
-
-/*****************************************************************************/
-
-/*
- * megamol::core::utility::log::Log::StreamTarget::StdOut
- */
-const std::shared_ptr<megamol::core::utility::log::Log::Target>
-megamol::core::utility::log::Log::StreamTarget::StdOut
-    = std::make_shared<megamol::core::utility::log::Log::StreamTarget>(std::cout);
-
-
-/*
- * megamol::core::utility::log::Log::StreamTarget::StdErr
- */
-const std::shared_ptr<megamol::core::utility::log::Log::Target>
-megamol::core::utility::log::Log::StreamTarget::StdErr
-    = std::make_shared<megamol::core::utility::log::Log::StreamTarget>(std::cerr);
-
-
-/*
- * megamol::core::utility::log::Log::StreamTarget::StreamTarget
- */
-megamol::core::utility::log::Log::StreamTarget::StreamTarget(std::ostream& stream, UINT level)
-        : Target(level) {
-    logger = spdlog::get("default_stream");
-    if (logger == nullptr) {
-        sink = std::make_shared<spdlog::sinks::ostream_sink_mt>(stream);
-        sink->set_pattern(std_pattern);
-        logger = std::make_shared<spdlog::logger>("default_stream", sink);
-    }
-}
-
-
-/*
- * megamol::core::utility::log::Log::StreamTarget::~StreamTarget
- */
-megamol::core::utility::log::Log::StreamTarget::~StreamTarget(void) {
-    // intentionally empty
-}
-
-
-/*
- * megamol::core::utility::log::Log::StreamTarget::Flush
- */
-void megamol::core::utility::log::Log::StreamTarget::Flush(void) {
-    logger->flush();
-}
-
-
-/*
- * megamol::core::utility::log::Log::StreamTarget::Msg
- */
-void megamol::core::utility::log::Log::StreamTarget::Msg(UINT level,
-        megamol::core::utility::log::Log::TimeStamp time, megamol::core::utility::log::Log::SourceID sid,
-        const char *msg) {
-    Msg(level, time, sid, std::string(msg));
-}
-
-
-/*
- * megamol::core::utility::log::Log::StreamTarget::Msg
- */
-void megamol::core::utility::log::Log::StreamTarget::Msg(UINT level, megamol::core::utility::log::Log::TimeStamp time,
-    megamol::core::utility::log::Log::SourceID sid, std::string const& msg) {
-    if ((level > this->Level())) return;
-    logger->info("{}|{}", level, msg);
-}
-
-/*****************************************************************************/
 
 /*
  * megamol::core::utility::log::Log::LEVEL_ALL
